created js builder for NewBookmarkListCard
