+++ source/PopNGo/PopNGo/PopNGo/wwwroot/js/ui/buildBookmarkListCard.js
@@ -1,49 +1,42 @@
 import { validateObject } from "../validation.js";
 
 /**
- * Takes in the the bookmark list card element and props and builds the card
+ * Takes in the the new bookmark list card element and props and builds the card
  * Props:
 {
-    bookmarkListName: String,
-    eventQuantity: Number,
-    onClick: Function,
- }
+    onClickCreateBookmarkList: (listName: String) => function,
+}
  * @function buildEventCard
- * @param {HTMLElement} eventCardElement 
- * @param {Object} props 
+ * @param {HTMLElement} bookmarkListCardElement 
+ * @param {Object?} props 
  */
-export const buildBookmarkListCard = (eventCardElement, props) => {
-    if (!validateBuildBookmarkListCardProps(props)) {
+export const buildBookmarkListCard = (bookmarkListCardElement, props) => {
+    if (!validateNewBuildBookmarkListCardProps(props)) {
         throw new Error('Invalid props');
     }
 
-    const { bookmarkListName, eventQuantity, onClick } = props;
+    // Wire up the create bookmark list button
+    const createBookmarkListButton = bookmarkListCardElement.querySelector('.saveNewBookmarkListButton');
 
-    // Set the bookmark list name
-    eventCardElement.querySelector('.bookmarkListCardTitleText').textContent = bookmarkListName;
-
-    // Set the event quantity
-    eventCardElement.querySelector('.bookmarkListCardQuantityText').textContent = eventQuantity;
-
-    // Add the event listener
-    eventCardElement.addEventListener('click', onClick);
+    createBookmarkListButton.addEventListener('click', () => {
+        const bookmarkListName = bookmarkListCardElement.querySelector('.newBookmarkListCardTitleInput').value;
+        props.onClick(bookmarkListName);
+    });
 }
 
 
 /**
- * Takes in the bookmark list card props and validates them
+ * Takes in the new bookmark list card props and validates them
  * @param {any} data
  * @returns {boolean}
  */
-export function validateBuildBookmarkListCardProps(data) {
+export function validateNewBuildBookmarkListCardProps(data) {
     if (data === undefined || data === null) {
-        return false;
+        return true;
     }
 
     const schema = {
-        bookmarkListName: x => typeof x === 'string',
-        eventQuantity: x => typeof x === 'number',
-        onClick: x => (typeof x === 'function' || x === undefined || x === null),
+        onClickCreateBookmarkList: x => (typeof x === 'function' || x === undefined || x === null),
     }
 
     return validateObject(data, schema).length === 0;
